convert notices and sitename to traits
