+++ app/Traits/Notices.php
@@ -1,0 +1,23 @@
+<?php
+
+namespace App\Traits;
+
+trait Notices
+{
+	/**
+	 * Returns notices from Advanced Custom Fields
+	 * @return array
+	 */
+	public function notices()
+	{
+		$notices = get_field('site_notices', 'options');
+		
+		if (isset($notices) && !empty($notices)) {
+			$notices = array_filter($notices, function($notice) use ($notices) {
+				return $notice['notice_active'] === true;
+			});
+		};
+
+		return $notices;
+	}
+}
