+++ app/Traits/SiteName.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace App\Traits;
+
+trait SiteName
+{
+	/**
+	 * Returns the name of the site
+	 * @return string
+	 */
+	public function siteName()
+	{
+		return get_bloginfo('name');
+	}
+}
